Move lines -> viewLines
This makes the naming consistent with the rest of the view parts where the
folder is named exactly after the view part

Fixes #227089
