+++ src/vs/editor/browser/viewParts/viewLines/domReadingContext.ts
