+++ src/vs/editor/browser/viewParts/viewLines/rangeUtil.ts
