+++ src/vs/editor/browser/viewParts/viewLines/viewLineOptions.ts
