+++ src/vs/editor/browser/viewParts/viewLines/viewLines.ts
